requirements.txt
python-telegram-bot[rate-limiter]==20.1
openai==1.3.3
tiktoken>=0.3.0
PyYAML==6.0
motor
decouple
tenacity==8.2.2
pydub~=0.25.1
pydantic==1.9.0
